+++ apps/@examples/modal/src/pages/_app.tsx
@@ -1,6 +1,7 @@
-import { useEffect } from 'react';
 import { DoptProvider } from '@dopt/react';
 import { useIdentifyUser } from '@/hooks';
+
+import { Example } from '@/pages/';
 
 export function App() {
   const userId = useIdentifyUser({
@@ -9,24 +10,13 @@ export function App() {
     inTrial: true,
   });
 
-  useEffect(() => {
-    if (userId) {
-      if (document.referrer !== '') {
-        window.parent.postMessage(
-          JSON.stringify({ userId }),
-          document.referrer
-        );
-      }
-    }
-  }, [userId]);
-
   return (
     <DoptProvider
       apiKey={import.meta.env.VITE_DOPT_BLOCKS_API_KEY}
       userId={userId}
       flowVersions={{ modal: 1 }}
     >
-      <>Modal Example</>
+      <Example />
     </DoptProvider>
   );
 }