+++ packages/@dopt/block-types/src/block/set.ts
@@ -15,4 +15,13 @@ export const SetProps = Type.Object(
 );
 
 export const Set = Type.Intersect([Base, SetProps], { $id: 'Group' });
+
+/**
+ * This type defines all the properties of a group (`set`) block.
+ * In addition to all the attributes of a regular step (or `model`) block,
+ * a group block also contains:
+ * - `size`: a number, the number of children blocks it encapsulates
+ * - `ordered`: a boolean, whether the group is ordered or unordered
+ * - `blocks`: an array, the array of all its child blocks
+ */
 export type Set = Static<typeof Set>;