+++ packages/@dopt/javascript/src/block.ts
@@ -13,21 +13,29 @@ import { blockStore } from './store';
 function resolveBlock(block: BlockType): BlockType {
   return blockStore.getState()[block.uid] || block;
 }
-
-interface Props {
+/**
+ * @internal
+ */
+export interface BlockProps {
   intent: ReturnType<typeof blocksApi>['blockIntent'];
   block: BlockType;
+  optimisticUpdates: boolean;
   fieldMap: Map<Field['sid'], Field> | null;
 }
 
-class Block {
-  private intent: Props['intent'];
-  private block: Props['block'];
-  private fieldMap: Props['fieldMap'];
+export class Block {
+  private intent: BlockProps['intent'];
+  private block: BlockProps['block'];
+  private fieldMap: BlockProps['fieldMap'];
+  private optimisticUpdates: BlockProps['optimisticUpdates'];
 
-  constructor({ block, intent, fieldMap }: Props) {
+  /**
+   * @internal
+   */
+  constructor({ block, intent, optimisticUpdates, fieldMap }: BlockProps) {
     this.intent = intent;
     this.block = block;
+    this.optimisticUpdates = optimisticUpdates;
     this.fieldMap = fieldMap;
   }
 
@@ -36,6 +44,11 @@ class Block {
     return this.intent({ uid, version, intent, goToUid });
   }
 
+  /**
+   * Gets the field (see {@link FieldType['value']}) with the `name` contained by this {@link Block}.
+   * If the {@link Block} does not have the field, the `defaultValue`
+   * is returned if provided. Otherwise, `null` is returned.
+   */
   getField<T extends FIELD_VALUE_UNION_TYPE>(
     name: string,
     defaultValue?: T
@@ -53,32 +66,90 @@ class Block {
       : null;
   }
 
+  /**
+   * Returns the up-to-date state of this {@link Block} instance.
+   *
+   * @returns The state of this instance.
+   */
   state(): BlockType['state'] {
     return resolveBlock(this.block).state;
   }
 
+  /**
+   * Complete this block. Will also update the state of others blocks within this flow (or the flow itself), as appropriate.
+   *
+   * @remarks
+   * It is often unnecessary to wait for this function to resolve / reject.
+   *
+   * @returns A promise which resolves when this block has been completed successfully and rejects otherwise.
+   */
   async complete() {
+    if (this.optimisticUpdates) {
+      const storedBlock = blockStore.getState()[this.block.uid];
+      if (storedBlock != null && storedBlock.type === ModelTypeConst) {
+        blockStore.setState({
+          [this.block.uid]: {
+            ...storedBlock,
+            state: { active: false, completed: true },
+          },
+        });
+      }
+    }
     return this._intent('complete');
   }
 
+  /**
+   * Go to the next block within an ordered group block. Returns an empty promise if this block instance isn't an ordered group.
+   *
+   * @remarks
+   * It is often unnecessary to wait for this function to resolve / reject.
+   *
+   * @returns A promise which resolves when this block has been completed successfully and rejects otherwise.
+   */
   async next() {
     if (this.block.type === SetTypeConst && this.block.ordered) {
       return this._intent('next');
     }
   }
 
+  /**
+   * Go to the previous block within an ordered group block. Returns an empty promise if this block instance isn't an ordered group.
+   *
+   * @remarks
+   * It is often unnecessary to wait for this function to resolve / reject.
+   *
+   * @returns A promise which resolves when this block has been completed successfully and rejects otherwise.
+   */
   async prev() {
     if (this.block.type === SetTypeConst && this.block.ordered) {
       return this._intent('prev');
     }
   }
 
-  async goTo(uid: string) {
+  /**
+   * Go to a specific block (by uid) within an ordered group block. Returns an empty promise if this block instance isn't an ordered group.
+   *
+   * @remarks
+   * It is often unnecessary to wait for this function to resolve / reject.
+   *
+   * @param uid The uid of the child step block.
+   *
+   * @returns A promise which resolves when this block has been completed successfully and rejects otherwise.
+   */
+  async goTo(uid: BlockType['uid']) {
     if (this.block.type === SetTypeConst && this.block.ordered) {
       return this._intent('goTo', uid);
     }
   }
 
+  /**
+   * Returns all completed (`completed: true`) children of a group ({@link SetType}) block.
+   *
+   * @remarks
+   * This returns an empty array if the instance is not of type {@link SetType}.
+   *
+   * @returns An array of {@link BlockType} objects. To access their classes, call `dopt.block` ({@link Dopt.block}).
+   */
   getCompleted(): BlockType[] {
     if (this.block.type === SetTypeConst) {
       return this.block.blocks
@@ -88,6 +159,15 @@ class Block {
 
     return [];
   }
+
+  /**
+   * Returns all uncompleted (`completed: false`) children of a group ({@link SetType}) block.
+   *
+   * @remarks
+   * This returns an empty array if the instance is not of type {@link SetType}.
+   *
+   * @returns An array of {@link BlockType} objects. To access their classes, call `dopt.block` ({@link Dopt.block}).
+   */
   getUncompleted(): BlockType[] {
     if (this.block.type === SetTypeConst) {
       return this.block.blocks
@@ -97,6 +177,15 @@ class Block {
 
     return [];
   }
+
+  /**
+   * Returns all active (`active: true`) children of a group ({@link SetType}) block.
+   *
+   * @remarks
+   * This returns an empty array if the instance is not of type {@link SetType}.
+   *
+   * @returns An array of {@link BlockType} objects. To access their classes, call `dopt.block` ({@link Dopt.block}).
+   */
   getActive(): BlockType[] {
     if (this.block.type === SetTypeConst) {
       return this.block.blocks
@@ -106,6 +195,15 @@ class Block {
 
     return [];
   }
+
+  /**
+   * Returns all inactive (`active: false`) children of a group ({@link SetType}) block.
+   *
+   * @remarks
+   * This returns an empty array if the instance is not of type {@link SetType}.
+   *
+   * @returns An array of {@link BlockType} objects. To access their classes, call `dopt.block` ({@link Dopt.block}).
+   */
   getInactive(): BlockType[] {
     if (this.block.type === SetTypeConst) {
       return this.block.blocks
@@ -115,9 +213,30 @@ class Block {
 
     return [];
   }
+
+  /**
+   * Subscribe to changes on this block.
+   *
+   * @example
+   * ```js
+   * const block = dopt.block("HNWvcT78tyTwygnbzU6SW");
+   * const unsubscribe = block.subscribe(async blockData => {
+   *  // access .state instead of .state()
+   *  // since block is an object of `BlockType`
+   *  if (block.state.completed) {
+   *     await showModal("Yay, you've completed your first step!");
+   *     unsubscribe();
+   *  }
+   * });
+   * ```
+   *
+   * @param listener
+   * The listener function is called with a {@link BlockType} object.
+   * You can use `dopt.block()` to access a {@link Block} instance instead.
+   *
+   * @returns A function which can be called to unsubscribe the listener.
+   */
   subscribe(listener: (block: BlockType) => void) {
-    blockStore.subscribe((blocks) => blocks[this.block.uid], listener);
+    return blockStore.subscribe((blocks) => blocks[this.block.uid], listener);
   }
 }
-
-export { Block };